changes from help

+++ src/pages/Home.jsx
@@ -44,7 +44,7 @@ const Home = () => {
         setUserInfo(response.data.user);
       }
     } catch (error) {
-      if (error.response.status === 401) {
+      if (error.response && error.response.status === 401) {
         localStorage.clear();
         navigate("/login");
       }
